fixed auth flow

+++ src/controllers/auth.controller.ts
@@ -16,159 +16,6 @@ const transporter = nodemailer.createTransport({
   },
 });
 
-const skills = [
-  "Web Development",
-  "WordPress",
-  "Python",
-  "Java",
-  "C++",
-  "App Development",
-  "Software Testing",
-  "Data Analytics",
-  "SQL",
-  "Cloud Computing",
-  "Cybersecurity",
-  "Blockchain",
-  "AI/ML",
-  "AR/VR",
-  "IoT",
-  "Automation",
-  "Game Development",
-  "Web Scraping",
-  "API Development",
-  "Chatbot Development",
-  "Graphic Design",
-  "UI/UX",
-  "Animation",
-  "Video Editing",
-  "3D Modeling",
-  "Logo Design",
-  "Infographics",
-  "Typography",
-  "NFT Art",
-  "Interior Design",
-  "Content Writing",
-  "Copywriting",
-  "Technical Writing",
-  "Ghostwriting",
-  "Resume Writing",
-  "Scriptwriting",
-  "Blogging",
-  "Research Writing",
-  "Translation",
-  "Transcription",
-  "Speech Writing",
-  "Social Media",
-  "SEO",
-  "Email Marketing",
-  "Ads Management",
-  "Affiliate Marketing",
-  "Influencer Marketing",
-  "PR",
-  "Market Research",
-  "Lead Generation",
-  "Growth Hacking",
-  "Accounting",
-  "Financial Analysis",
-  "Stock Trading",
-  "Cryptocurrency",
-  "Tax Filing",
-  "Budgeting",
-  "Crowdfunding",
-  "Business Valuation",
-  "Investment Analysis",
-  "Risk Management",
-  "Public Speaking",
-  "Negotiation",
-  "Conflict Resolution",
-  "Time Management",
-  "Leadership",
-  "Networking",
-  "Emotional Intelligence",
-  "Personal Branding",
-  "Interviewing",
-  "Problem-Solving",
-  "Online Tutoring",
-  "Language Teaching",
-  "Music Lessons",
-  "Fitness Training",
-  "Life Coaching",
-  "Career Counseling",
-  "Exam Coaching",
-  "Yoga",
-  "Personal Development",
-  "Skill Training",
-  "Virtual Assistance",
-  "Data Entry",
-  "Email Management",
-  "Customer Support",
-  "Travel Planning",
-  "Project Management",
-  "Event Planning",
-  "Document Formatting",
-  "CRM Management",
-  "E-commerce",
-  "Dropshipping",
-  "Product Listing",
-  "Sales Funnels",
-  "Print-on-Demand",
-  "B2B Sales",
-  "Customer Retention",
-  "Online Courses",
-  "Subscription Business",
-  "Retail Management",
-  "Podcasting",
-  "Voiceover",
-  "Photography",
-  "Freelance Writing",
-  "Handmade Crafts",
-  "Car Maintenance",
-  "Home Repair",
-  "Cooking",
-  "Nutrition",
-  "First Aid",
-  "Emergency Preparedness",
-  "Gardening",
-  "Public Transport Navigation",
-  "Apartment Hunting",
-  "Resume Optimization",
-  "Personal Finance",
-  "Stress Management",
-  "Meditation",
-  "Relationship Building",
-  "Workplace Communication",
-  "Professional Dressing",
-  "Job Search",
-  "Business Proposal",
-  "Legal Knowledge",
-  "Debt Management",
-  "Stock Photography",
-  "Virtual Reality Content",
-  "3D Printing",
-  "Ethical Hacking",
-  "Cloud Security",
-  "Copyediting",
-  "Data Visualization",
-  "Business Consulting",
-  "HR Management",
-  "Podcast Editing",
-  "Mobile Repair",
-  "Digital Illustration",
-  "App Monetization",
-  "Video Marketing",
-  "Email Copywriting",
-  "Digital Fundraising",
-  "Dance Choreography",
-  "DIY Home Decor",
-  "Resume Designing",
-  "Smart Home Setup",
-  "Google Analytics",
-  "Social Media Ads",
-  "Public Relations Writing",
-  "Proofreading",
-  "Voice Modulation",
-];
-
 const sendOtp = async (email: string, otp: string) => {
   try {
     const mailOptions = {
@@ -200,11 +47,12 @@ export const register = async (
   next: NextFunction,
 ): Promise<any> => {
   try {
-    const { email } = req.body;
+    const { email, username } = req.body;
 
     const existingUser = await prisma.user.findFirst({
       where: {
         email,
+        username
       },
     });
 
@@ -265,17 +113,24 @@ export const verifyOtp = async (
       endYear,
     } = req.body;
 
-    // Delete expired OTPs
+    const user = await prisma.user.findUnique({
+      where: { email },
+    });
+
+    if (!user) {
+      return res.status(400).json({ error: "User not found" });
+    }
+
     await prisma.otp.deleteMany({
       where: {
-        email,
-        expiresAt: { lt: new Date() }, // Delete expired OTPs before verification
+        userId: user.id,
+        expiresAt: { lt: new Date() },
       },
     });
 
     // Check OTP
     const otpRecord = await prisma.otp.findFirst({
-      where: { email, code: otp },
+      where: { userId: user.id, code: otp },
       orderBy: { createdAt: "desc" },
     });
 
@@ -287,19 +142,15 @@ export const verifyOtp = async (
       return res.status(400).json({ error: "OTP expired. Request a new OTP" });
     }
 
-    // Delete OTP after verification
     await prisma.otp.delete({ where: { id: otpRecord.id } });
 
-    // Check if user is already verified
     const existingUser = await prisma.user.findUnique({ where: { email } });
     if (existingUser?.verified) {
       return res.status(400).json({ error: "Email already registered" });
     }
 
-    // Hash Password
     const hashedPassword = await bcrypt.hash(password, 10);
 
-    // Create or update user
     const updatedUser = await prisma.user.upsert({
       where: { email },
       update: {
@@ -320,7 +171,7 @@ export const verifyOtp = async (
         degree,
         startYear,
         endYear,
-        verified: true, // Mark as verified after OTP
+        verified: true,
       },
       create: {
         email,
@@ -345,7 +196,6 @@ export const verifyOtp = async (
       },
     });
 
-    // Generate Token
     const token = generateToken(updatedUser.id);
     res.status(201).json({ token, user: updatedUser });
   } catch (error) {
